+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - If no translator is specified, a default translator is created automatically.
 - Rule enforcing that passwords don't contain words from a dictionary.
 - Converters for changing the case of letters in words.
+- Pspell can be used as word list for dictionaries.
 
 ### Changed
 - Translator no longer needs to be specified when enforcing rules.